+++ allauth/socialaccount/providers/oauth2/views.py
@@ -27,6 +27,7 @@ from ..base import AuthAction, AuthError
 
 class OAuth2Adapter(object):
     expires_in_key = 'expires_in'
+    client_cls = OAuth2Client
     supports_state = True
     redirect_uri_protocol = None
     access_token_method = 'POST'
@@ -61,6 +62,9 @@ class OAuth2Adapter(object):
                 seconds=int(expires_in))
         return token
 
+    def get_access_token_data(self, request, app, client):
+        code = get_request_param(self.request, "code")
+        return client.get_access_token(code)
 
 class OAuth2View(object):
     @classmethod
@@ -79,14 +83,20 @@ class OAuth2View(object):
         callback_url = self.adapter.get_callback_url(request, app)
         provider = self.adapter.get_provider()
         scope = provider.get_scope(request)
-        client = OAuth2Client(self.request, app.client_id, app.secret,
-                              self.adapter.access_token_method,
-                              self.adapter.access_token_url,
-                              callback_url,
-                              scope,
-                              scope_delimiter=self.adapter.scope_delimiter,
-                              headers=self.adapter.headers,
-                              basic_auth=self.adapter.basic_auth)
+        client = self.adapter.client_cls(
+            self.request,
+            app.client_id,
+            app.secret,
+            self.adapter.access_token_method,
+            self.adapter.access_token_url,
+            callback_url,
+            scope,
+            key=app.key,
+            cert=app.cert,
+            scope_delimiter=self.adapter.scope_delimiter,
+            headers=self.adapter.headers,
+            basic_auth=self.adapter.basic_auth,
+        )
         return client
 
 
@@ -111,7 +121,9 @@ class OAuth2LoginView(OAuth2View):
 
 class OAuth2CallbackView(OAuth2View):
     def dispatch(self, request, *args, **kwargs):
-        if 'error' in request.GET or 'code' not in request.GET:
+        auth_error = get_request_param(request, "error")
+        code = get_request_param(request, "code")
+        if auth_error or not code:
             # Distinguish cancel from error
             auth_error = request.GET.get('error', None)
             if auth_error == self.adapter.login_cancelled_error:
@@ -123,21 +135,21 @@ class OAuth2CallbackView(OAuth2View):
                 self.adapter.provider_id,
                 error=error)
         app = self.adapter.get_provider().get_app(self.request)
-        client = self.get_client(request, app)
+        client = self.get_client(self.request, app)
         try:
-            access_token = client.get_access_token(request.GET['code'])
-            token = self.adapter.parse_token(access_token)
+            token_data = self.adapter.get_access_token_data(
+                self.request, app=app, client=client
+            )
+            token = self.adapter.parse_token(data=token_data)
             token.app = app
             login = self.adapter.complete_login(request,
                                                 app,
                                                 token,
-                                                response=access_token)
+                                                response=token_data)
             login.token = token
+            state = get_request_param(request, "state")
             if self.adapter.supports_state:
-                login.state = SocialLogin \
-                    .verify_and_unstash_state(
-                        request,
-                        get_request_param(request, 'state'))
+                login.state = SocialLogin.verify_and_unstash_state(request, state)
             else:
                 login.state = SocialLogin.unstash_state(request)
             return complete_social_login(request, login)